split updateProps
